Реализовать сервис статистики (Реструктуризация)
