+++ ewm-main-service/src/main/java/ru/practicum/explorewithme/main/EwmMainApp.java
