--- ewm-main/src/main/resources/application.properties
@@ -1,17 +1,0 @@
-server.port=9090
-
-spring.jpa.hibernate.ddl-auto=none
-spring.jpa.properties.hibernate.format_sql=true
-spring.sql.init.mode=always
-
-logging.level.org.springframework.orm.jpa=INFO
-logging.level.org.springframework.transaction=INFO
-logging.level.org.springframework.transaction.interceptor=TRACE
-logging.level.org.springframework.orm.jpa.JpaTransactionManager=DEBUG
-logging.level.org.hibernate.SQL=DEBUG
-logging.level.org.hibernate.orm.jdbc.bind=trace
-
-spring.datasource.driverClassName=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://localhost:5432/ewm-db
-spring.datasource.username=dbuser
-spring.datasource.password=dbuser
