+++ ewm-stats-server/stat-client/src/main/java/ru/practicum/explorewithme/statclient/ClientStat.java
@@ -1,0 +1,50 @@
+package ru.practicum.explorewithme.statclient;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.http.ResponseEntity;
+import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
+import org.springframework.stereotype.Service;
+import org.springframework.web.util.DefaultUriBuilderFactory;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import java.util.Map;
+
+@Service
+public class ClientStat extends ClientBase {
+
+    private final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
+
+    @Autowired
+    public ClientStat(@Value("${stat-server.url}") String serverUrl, RestTemplateBuilder builder) {
+        super(
+                builder
+                    .uriTemplateHandler(new DefaultUriBuilderFactory(serverUrl))
+                    .requestFactory(() -> new HttpComponentsClientHttpRequestFactory())
+                    .build()
+        );
+    }
+
+    public ResponseEntity<Object> getStats(LocalDateTime start, LocalDateTime end, String[] uris, Boolean unique) {
+        Map<String, Object> parameters = Map.of(
+                "start", start.format(formatter),
+                "end", end.format(formatter),
+                "uris", uris == null ? "" : String.join(",", uris),
+                "unique", unique
+        );
+        return get("/stats?start={start}&end={end}&uris={uris}&unique={unique}", parameters);
+    }
+
+    public ResponseEntity<Object> save(String app, String uri, String ip) {
+        Map<String, Object> parameters = Map.of(
+                "app", app,
+                "uri", uri,
+                "ip", ip,
+                "timestamp", LocalDateTime.now().format(formatter)
+        );
+        return post("/hit", parameters);
+    }
+
+}
