+++ ewm-stats-server/stat-client/src/main/java/ru/practicum/explorewithme/statclient/ExploreWithMeStatClientApp.java
@@ -1,0 +1,11 @@
+package ru.practicum.explorewithme.statclient;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ExploreWithMeStatClientApp {
+    public static void main(String[] args) {
+        SpringApplication.run(ExploreWithMeStatClientApp.class, args);
+    }
+}
