+++ ewm-stats-server/stat-client/src/main/resources/application.properties
@@ -1,0 +1,5 @@
+logging.level.org.springframework.web.client.RestTemplate=DEBUG
+logging.level.org.apache.http=DEBUG
+logging.level.httpclient.wire=DEBUG
+
+stat-server.url=http://localhost:9090
