+++ ewm-stats-server/stat-dto/src/main/java/ru/practicum/explorewithme/statdto/StatDto.java
@@ -1,0 +1,12 @@
+package ru.practicum.explorewithme.statdto;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class StatDto {
+    private String app;
+    private String uri;
+    private Long hits;
+}
