+++ ewm-stats-server/stat-server/src/main/java/ru/practicum/explorewithme/statserver/EwmStatServerApp.java
@@ -1,0 +1,11 @@
+package ru.practicum.explorewithme.statserver;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EwmStatServerApp {
+    public static void main(String[] args) {
+        SpringApplication.run(EwmStatServerApp.class, args);
+    }
+}
