+++ ewm-stats-server/stat-server/src/main/java/ru/practicum/explorewithme/statserver/model/Stat.java
@@ -1,0 +1,33 @@
+package ru.practicum.explorewithme.statserver.model;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "stat", schema = "public")
+public class Stat {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String app;
+
+    private String uri;
+
+    private String ip;
+
+    @Column(name = "stime")
+    private LocalDateTime timestamp;
+}
