+++ ewm-stats-server/stat-server/src/main/java/ru/practicum/explorewithme/statserver/model/StatWithHits.java
@@ -1,0 +1,11 @@
+package ru.practicum.explorewithme.statserver.model;
+
+public interface StatWithHits {
+
+    String getApp();
+
+    String getUri();
+
+    Long getHits();
+
+}
