+++ ewm-stats-server/stat-server/src/main/resources/schema.sql
@@ -1,0 +1,7 @@
+CREATE TABLE IF NOT EXISTS stat (
+  id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
+  app VARCHAR(255) NOT NULL,
+  uri VARCHAR(255) NOT NULL,
+  ip VARCHAR(15) NOT NULL,
+  stime TIMESTAMP NOT NULL
+);
