--- ewm-stats/client/src/main/java/ru/practicum/explorewithme/stats/client/EwmClientApp.java
@@ -1,11 +1,0 @@
-package ru.practicum.explorewithme.stats.client;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EwmClientApp {
-    public static void main(String[] args) {
-        SpringApplication.run(EwmClientApp.class, args);
-    }
-}
