--- ewm-stats/client/src/main/resources/application-test.properties
@@ -1,1 +1,0 @@
-stat-server.url=http://localhost:9090
