--- ewm-stats/client/src/main/resources/application.properties
@@ -1,5 +1,0 @@
-logging.level.org.springframework.web.client.RestTemplate = DEBUG
-logging.level.org.apache.http = DEBUG
-logging.level.httpclient.wire = DEBUG
-
-stats-server.url = http://localhost:9090
