--- ewm-stats/dto/src/main/java/ru/practicum/explorewithme/stats/dto/HitDto.java
@@ -1,13 +1,0 @@
-package ru.practicum.explorewithme.stats.dto;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class HitDto {
-    private String app;
-    private String uri;
-    private String ip;
-    private String timestamp;
-}
