--- ewm-stats/server/src/main/java/ru/practicum/explorewithme/stats/server/EwmServerApp.java
@@ -1,11 +1,0 @@
-package ru.practicum.explorewithme.stats.server;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EwmServerApp {
-    public static void main(String[] args) {
-        SpringApplication.run(EwmServerApp.class, args);
-    }
-}
