--- ewm-stats/server/src/main/java/ru/practicum/explorewithme/stats/server/controller/StatisticsController.java
@@ -1,44 +1,0 @@
-package ru.practicum.explorewithme.stats.server.controller;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.format.annotation.DateTimeFormat;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-import ru.practicum.explorewithme.stats.dto.HitDto;
-import ru.practicum.explorewithme.stats.dto.StatDto;
-import ru.practicum.explorewithme.stats.server.service.StatisticsService;
-
-import java.time.LocalDateTime;
-import java.util.Collection;
-import java.util.List;
-
-@RestController
-@RequiredArgsConstructor
-@Slf4j
-public class StatisticsController {
-
-    private final StatisticsService statsService;
-
-    @PostMapping("/hit")
-    @ResponseStatus(HttpStatus.CREATED)
-    public void hit(@RequestBody HitDto hitDto) {
-        log.info("Получен запрос на сохранение статистики посещений с данными: {}", hitDto);
-        statsService.saveHit(hitDto);
-    }
-
-    @GetMapping("/stats")
-    public Collection<StatDto> getStats(@RequestParam @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime start,
-                                        @RequestParam @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime end,
-                                        @RequestParam(value = "uris", required = false) List<String> uris,
-                                        @RequestParam(value = "unique", defaultValue = "false") Boolean unique) {
-        log.info("Получен запрос на получение статистики посещений c параметрами: start={}, end={}, uris={}, unique={}", start, end, uris, unique);
-        return statsService.getStats(start, end, uris, unique);
-    }
-
-}
