--- ewm-stats/server/src/main/java/ru/practicum/explorewithme/stats/server/model/StatisticWithHits.java
@@ -1,10 +1,0 @@
-package ru.practicum.explorewithme.stats.server.model;
-
-public interface StatisticWithHits {
-
-    String getApp();
-
-    String getUri();
-
-    Long getHits();
-}
